Added support for SC 9.0.2

+++ src/Contrib.Sitecore.ContentSearch.TikaOnDotnet/ComputedFields/MediaItemTikaOnDotNetTextExtractor.cs
@@ -71,7 +71,10 @@ namespace Contrib.Sitecore.ContentSearch.TikaOnDotnet.ComputedFields
             }
             finally
             {
-                File.Delete(tempFilePath);
+                if (File.Exists(tempFilePath))
+                {
+                    File.Delete(tempFilePath);
+                }
             }
         }
